+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/InAppWebView/FlutterWebView.java
@@ -54,6 +54,13 @@ public class FlutterWebView implements PlatformView, MethodCallHandler  {
     InAppWebViewOptions options = new InAppWebViewOptions();
     options.parse(initialOptions);
 
+    if (Shared.activity == null) {
+      Log.e(LOG_TAG, "\n\n\nERROR: Shared.activity is null!!!\n\n" +
+              "You need to upgrade your Flutter project to use the new Java Embedding API:\n\n" +
+              "- Take a look at the \"IMPORTANT Note for Android\" section here: https://github.com/pichillilorenzo/flutter_inappwebview#important-note-for-android\n" +
+              "- See the official wiki here: https://github.com/flutter/flutter/wiki/Upgrading-pre-1.12-Android-projects\n\n\n");
+    }
+
     webView = new InAppWebView(Shared.activity, this, id, options, contextMenu, containerView);
     displayListenerProxy.onPostWebViewInitialization(displayManager);