+++ example/lib/main.dart
@@ -6,13 +6,15 @@ import 'package:flutter_inappwebview_example/chrome_safari_browser_example.scree
 import 'package:flutter_inappwebview_example/headless_in_app_webview.screen.dart';
 import 'package:flutter_inappwebview_example/in_app_webiew_example.screen.dart';
 import 'package:flutter_inappwebview_example/in_app_browser_example.screen.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 // InAppLocalhostServer localhostServer = new InAppLocalhostServer();
 
 Future main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  // await Permission.camera.request();
   // await localhostServer.start();
-  runApp(new MyApp());
+  runApp(MyApp());
 }
 
 Drawer myDrawer({@required BuildContext context}) {